app/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
databricks-sdk>=0.81.0
psycopg[binary]>=3.0
